requirements.txt
Flask
Flask-SQLAlchemy
google-generativeai
python-dotenv
pandas
SQLAlchemy
gunicorn
faiss-cpu
numpy
requests
scipy
Werkzeug
psycopg2-binary
python-dotenv
# sentence-transformers